A Config file for the app in .ini format.
A Settings Class that when initialised will read the ini file and allow access to all of the values. It also has a method called "updateSettings" that will re-read the ini file.

A Messenger class. This class is to print out messages and deal with errors on the application all in the same way. The "printMessage" method will probably end up producing a GUI message box on screen on which the user presses ok.